requirements.txt
flask
opencv-python
Pillow
prometheus_client
pytube
yt-dlp
